[LCR-206] refactor: controller 패키지명을 externalApi로 변경
