+++ popi-member-service/src/main/java/com/lgcns/externalApi/MemberController.java
@@ -1,4 +1,4 @@
-package com.lgcns.controller;
+package com.lgcns.externalApi;
 
 import com.lgcns.dto.response.MemberInfoResponse;
 import com.lgcns.service.MemberService;
